Remove server code
