+++ Makefile.toml
@@ -28,14 +28,16 @@ dependencies = ["build-setup"]
 [tasks.build]
 description = "Build development binaries in debug mode"
 command = "cargo"
-args = ["build", "--features=dev", "--package=omni-led", "--package=audio", "--package=clock",
-    "--package=images", "--package=media", "--package=system", "--package=weather"]
+args = ["build", "--features=dev", "--package=omni-led", "--package=clock" ]
+# args = ["build", "--features=dev", "--package=omni-led", "--package=audio", "--package=clock",
+#     "--package=images", "--package=media", "--package=system", "--package=weather"]
 
 [tasks.build-release]
 description = "Build development binaries in release mode"
 command = "cargo"
-args = ["build", "--release", "--features=dev", "--package=omni-led", "--package=audio",
-    "--package=clock", "--package=images", "--package=media", "--package=system", "--package=weather"]
+args = ["build", "--release", "--features=dev", "--package=omni-led", "--package=clock" ]
+# args = ["build", "--release", "--features=dev", "--package=omni-led", "--package=audio",
+#     "--package=clock", "--package=images", "--package=media", "--package=system", "--package=weather"]
 
 [tasks.run]
 description = "Run development binaries in debug mode"
