+++ config/applications.lua
@@ -1,39 +1,39 @@
 load_plugin {
     path = get_default_plugin_path('clock'),
-    args = { '--address', SERVER.Address },
+    args = { },
 }
 
-load_app {
-    path = get_default_path('audio'),
-    args = { '--address', SERVER.Address },
-}
-
-load_app {
-    path = get_default_path('media'),
-    args = {
-        '--address', SERVER.Address,
-        '--mode', 'individual',
-        '--map', 'Spotify.exe=SPOTIFY',
-    },
-}
-
-load_app {
-    path = get_default_path('system'),
-    args = {
-        '--address', SERVER.Address,
-        '--interval', '2sec',
-    },
-}
-
-load_app {
-    path = get_default_path('weather'),
-    args = {
-        '--address', SERVER.Address,
-        '--interval', '10min',
-        '--wind-speed-unit', 'm/s',
-        '--temperature-unit', 'Celsius',
-        'in', 'Warsaw',
-        '--country-code', 'PL',
-        '--administrative', 'Mazovia',
-    }
-}
+-- load_app {
+--     path = get_default_path('audio'),
+--     args = { '--address', SERVER.Address },
+-- }
+--
+-- load_app {
+--     path = get_default_path('media'),
+--     args = {
+--         '--address', SERVER.Address,
+--         '--mode', 'individual',
+--         '--map', 'Spotify.exe=SPOTIFY',
+--     },
+-- }
+--
+-- load_app {
+--     path = get_default_path('system'),
+--     args = {
+--         '--address', SERVER.Address,
+--         '--interval', '2sec',
+--     },
+-- }
+--
+-- load_app {
+--     path = get_default_path('weather'),
+--     args = {
+--         '--address', SERVER.Address,
+--         '--interval', '10min',
+--         '--wind-speed-unit', 'm/s',
+--         '--temperature-unit', 'Celsius',
+--         'in', 'Warsaw',
+--         '--country-code', 'PL',
+--         '--administrative', 'Mazovia',
+--     }
+-- }
