+++ config/settings.lua
@@ -5,6 +5,5 @@ Settings {
     log_level = LevelFilter.Debug,
     keyboard_ticks_repeat_delay = 2,
     keyboard_ticks_repeat_rate = 2,
-    server_port = 0,
     update_interval = Duration.from_millis(100),
 }
