+++ omni-led-api/Cargo.toml
@@ -1,19 +1,13 @@
 [package]
 name = "omni-led-api"
 version = "0.9.1"
-authors = ["Michał Bałabanow <m.balabanow@gmail.com>"]
-license = "GPL-3.0-only"
 edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 image = "0.25"
 log = { version = "0.4", features = ["std"] }
 prost = "0.14"
-tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
-tonic = "0.14"
-tonic-prost = "0.14"
 
 [build-dependencies]
-tonic-prost-build = "0.14"
+bindgen = "0.71.0"
+prost-build = "0.14"
