--- omni-led-api/proto/plugin.proto
@@ -1,80 +1,0 @@
-syntax = 'proto3';
-
-package plugin;
-
-service Plugin {
-  rpc Event(EventData) returns (EventResponse);
-  rpc Log(LogData) returns (LogResponse);
-}
-
-message EventData {
-  string name = 1;
-  Table fields = 2;
-}
-
-message EventResponse {}
-
-message Field {
-  oneof field {
-    None f_none = 8;
-    bool f_bool = 1;
-    int64 f_integer = 2;
-    double f_float = 3;
-    string f_string = 4;
-    Array f_array = 5;
-    Table f_table = 6;
-    ImageData f_image_data = 7;
-  }
-}
-
-message Array {
-  repeated Field items = 1;
-}
-
-message Table {
-  map<string, Field> items = 1;
-}
-
-message ImageData {
-  ImageFormat format = 1;
-  bytes data = 2;
-}
-
-enum ImageFormat {
-  IMAGE_FORMAT_UNKNOWN = 0;
-  IMAGE_FORMAT_PNG = 1;
-  IMAGE_FORMAT_JPEG = 2;
-  IMAGE_FORMAT_GIF = 3;
-  IMAGE_FORMAT_WEBP = 4;
-  IMAGE_FORMAT_PNM = 5;
-  IMAGE_FORMAT_TIFF = 6;
-  IMAGE_FORMAT_TGA = 7;
-  IMAGE_FORMAT_DDS = 8;
-  IMAGE_FORMAT_BMP = 9;
-  IMAGE_FORMAT_ICO = 10;
-  IMAGE_FORMAT_HDR = 11;
-  IMAGE_FORMAT_OPEN_EXR = 12;
-  IMAGE_FORMAT_FARBFELD = 13;
-  IMAGE_FORMAT_AVIF = 14;
-  IMAGE_FORMAT_QOI = 15;
-  IMAGE_FORMAT_PCX = 16 [deprecated = true];
-}
-
-message LogData {
-  LogLevel log_level = 1;
-  string location = 2;
-  string message = 3;
-}
-
-enum LogLevel {
-  LOG_LEVEL_UNKNOWN = 0;
-  LOG_LEVEL_ERROR = 1;
-  LOG_LEVEL_WARN = 2;
-  LOG_LEVEL_INFO = 3;
-  LOG_LEVEL_DEBUG = 4;
-  LOG_LEVEL_TRACE = 5;
-}
-
-message LogResponse {}
-
-message None {}
