+++ omni-led-lib/Cargo.toml
@@ -27,11 +27,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 softbuffer = "0.4"
 omni-led-api = { path = "../omni-led-api" }
-omni-led-new-api = { path = "../omni-led-new-api" }
 omni-led-derive = { path = "../omni-led-derive", features = ["from-lua-value", "lua-enum", "unique-user-data"] }
-tokio = { version = "1", features = ["fs", "macros", "rt-multi-thread"] }
-tokio-stream = "0.1"
-tonic = "0.14"
 tray-icon = "0.22"
 ureq = { version = "3.1", features = ["json"] }
 winit = "0.30"
