+++ omni-led-lib/src/lib.rs
@@ -8,6 +8,5 @@ pub mod logging;
 pub mod renderer;
 pub mod script_handler;
 pub mod semaphore;
-pub mod server;
 pub mod settings;
 pub mod ui;
