--- omni-led-lib/src/server/mod.rs
@@ -1,1 +1,0 @@
-pub mod server;
