--- omni-led-lib/src/server/server.rs
@@ -1,155 +1,0 @@
-use log::{debug, error, log};
-use mlua::{Lua, UserData, UserDataFields};
-use omni_led_api::plugin::Plugin;
-use omni_led_api::types::{EventData, EventResponse, LogData, LogLevel, LogResponse};
-use serde::Serialize;
-use std::sync::{Arc, Mutex};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-use tokio::net::TcpListener;
-use tokio::runtime::Runtime;
-use tokio::sync::oneshot;
-use tonic::transport::Server;
-use tonic::{Code, Request, Response, Status};
-
-use crate::common::user_data::UserDataRef;
-use crate::constants::constants::Constants;
-use crate::events;
-use crate::events::event_queue::EventQueue;
-use crate::settings::settings::Settings;
-
-pub struct PluginServer {
-    event_queue: Arc<Mutex<EventQueue>>,
-}
-
-impl PluginServer {
-    pub fn load(lua: &Lua, rt: &Runtime) -> oneshot::Sender<()> {
-        const LOCALHOST: &str = "127.0.0.1";
-
-        let settings = UserDataRef::<Settings>::load(lua);
-
-        let port: u16 = settings.get().server_port;
-
-        let listener = rt
-            .block_on(TcpListener::bind(format!("{LOCALHOST}:{port}")))
-            .unwrap();
-
-        let address = listener.local_addr().unwrap();
-        let bound_port = address.port();
-
-        let (tx, rx) = oneshot::channel();
-        rt.spawn(
-            Server::builder()
-                .add_service(
-                    omni_led_api::types::plugin_server::PluginServer::new(Self::new())
-                        .max_decoding_message_size(64 * 1024 * 1024)
-                        .max_encoding_message_size(64 * 1024 * 1024),
-                )
-                .serve_with_incoming_shutdown(
-                    tokio_stream::wrappers::TcpListenerStream::new(listener),
-                    async {
-                        _ = rx.await;
-                        debug!("Server shutting down")
-                    },
-                ),
-        );
-
-        let address = format!("{LOCALHOST}:{bound_port}");
-        let timestamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_else(|err| {
-                error!("Failed to get unix epoch time: {}", err);
-                Duration::ZERO
-            })
-            .as_millis() as u64;
-        let info = ServerInfo {
-            address,
-            ip: format!("{LOCALHOST}"),
-            port: bound_port,
-            timestamp,
-        };
-
-        let constants = UserDataRef::<Constants>::load(lua);
-        std::fs::write(
-            constants.get().data_dir.join("server.json"),
-            serde_json::to_string_pretty(&info).unwrap(),
-        )
-        .unwrap();
-
-        lua.globals().set("SERVER", info).unwrap();
-
-        tx
-    }
-
-    fn new() -> Self {
-        Self {
-            event_queue: EventQueue::instance(),
-        }
-    }
-}
-
-#[derive(Clone, Serialize)]
-struct ServerInfo {
-    address: String,
-    ip: String,
-    port: u16,
-    timestamp: u64,
-}
-
-impl UserData for ServerInfo {
-    fn add_fields<F: UserDataFields<Self>>(fields: &mut F) {
-        fields.add_field_method_get("Address", |_, info| Ok(info.address.clone()));
-        fields.add_field_method_get("Ip", |_, info| Ok(info.ip.clone()));
-        fields.add_field_method_get("Port", |_, info| Ok(info.port));
-        fields.add_field_method_get("Timestamp", |_, info| Ok(info.timestamp));
-    }
-}
-
-#[tonic::async_trait]
-impl omni_led_api::types::plugin_server::Plugin for PluginServer {
-    async fn event(
-        &self,
-        mut request: Request<EventData>,
-    ) -> Result<Response<EventResponse>, Status> {
-        let event = request.get_mut();
-
-        if !Plugin::is_valid_identifier(&event.name) {
-            return Err(Status::new(Code::InvalidArgument, "Invalid event name"));
-        }
-
-        let mut name = String::new();
-        std::mem::swap(&mut name, &mut event.name);
-
-        let mut fields = None;
-        std::mem::swap(&mut fields, &mut event.fields);
-
-        let fields = match fields {
-            Some(fields) => fields,
-            None => std::unreachable!(
-                "This has type Option<Table> due the code generator, field is required in proto definition"
-            ),
-        };
-
-        self.event_queue
-            .lock()
-            .unwrap()
-            .push(events::event_queue::Event::Application((name, fields)));
-
-        Ok(Response::new(EventResponse {}))
-    }
-
-    async fn log(&self, request: Request<LogData>) -> Result<Response<LogResponse>, Status> {
-        let data = request.get_ref();
-
-        let location = format!("plugin::{}", data.location);
-        match data.log_level() {
-            LogLevel::Unknown => {
-                error!(target: &location, "Received unknown log level. Original log message: '{}'", data.message)
-            }
-            level => {
-                log!(target: &location, level.into(), "{}", data.message);
-            }
-        };
-
-        Ok(Response::new(LogResponse {}))
-    }
-}
