--- omni-led-new-api/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod c_api;
-pub mod logging;
-pub mod plugin;
-pub mod rust_api;
-pub mod types;
