+++ omni-led/Cargo.toml
@@ -11,7 +11,6 @@ log = { version = "0.4", features = ["std"] }
 log4rs = "1.4"
 mlua = "0.11.4"
 omni-led-lib = { path = "../omni-led-lib" }
-tokio = { version = "1", features = ["fs", "macros", "rt-multi-thread"] }
 
 [features]
 dev = ["omni-led-lib/dev"]
