+++ omni-led/src/main.rs
@@ -16,14 +16,13 @@ use omni_led_lib::{
     keyboard::keyboard::process_events,
     logging::logger::Log,
     script_handler::script_handler::ScriptHandler,
-    server::server::PluginServer,
     settings::settings::Settings,
     ui::event::Event,
     ui::handler::{HandlerBuilder, PROXY},
 };
 use std::sync;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 mod logging;
 
@@ -36,11 +35,6 @@ fn main() {
     let (ready_tx, ready_rx) = sync::mpsc::channel();
 
     let scripting_thread = std::thread::spawn(move || {
-        let rt = tokio::runtime::Builder::new_multi_thread()
-            .enable_all()
-            .build()
-            .unwrap();
-
         let init_begin = Instant::now();
 
         let lua = Lua::new();
@@ -63,7 +57,6 @@ fn main() {
         let settings_config = read_config(&lua, ConfigType::Settings).unwrap();
 
         Settings::load(&lua, settings_config);
-        let server_shutdown_tx = PluginServer::load(&lua, &rt);
         let mut dispatcher = Dispatcher::load(&lua);
         Events::load(&lua);
         Shortcuts::load(&lua);
@@ -85,9 +78,6 @@ fn main() {
             let mut script_handler = UserDataRef::<ScriptHandler>::load(&lua);
             script_handler.get_mut().update(&lua, interval).unwrap();
         });
-
-        server_shutdown_tx.send(()).unwrap();
-        rt.shutdown_timeout(Duration::from_secs(1));
     });
 
     let keyboard_thread = std::thread::spawn(|| process_events(&RUNNING));
